+++ NINA.Image/ImageData/StarDetectionAnalysis.cs
@@ -12,6 +12,7 @@
 
 #endregion "copyright"
 
+using NINA.Core.Enum;
 using NINA.Core.Utility;
 using NINA.Image.ImageAnalysis;
 using NINA.Image.Interfaces;
@@ -24,6 +25,9 @@ namespace NINA.Image.ImageData {
         private double _fwhm = double.NaN;
         private double _eccentricity = double.NaN;
         private double _hfrStDev = double.NaN;
+        private StarMeasurementUnit _hfrUnit = StarMeasurementUnit.Pixels;
+        private StarMeasurementUnit _fwhmUnit = StarMeasurementUnit.Arcseconds;
+        private StarMeasurementUnit _hfrStDevUnit = StarMeasurementUnit.Pixels;
         private int _detectedStars = -1;
         private List<DetectedStar> _starList = new List<DetectedStar>();
 
@@ -55,6 +59,30 @@ namespace NINA.Image.ImageData {
             get => this._hfrStDev;
             set {
                 this._hfrStDev = value;
+                this.RaisePropertyChanged();
+            }
+        }
+
+        public StarMeasurementUnit HFRUnit {
+            get => this._hfrUnit;
+            set {
+                this._hfrUnit = value;
+                this.RaisePropertyChanged();
+            }
+        }
+
+        public StarMeasurementUnit FWHMUnit {
+            get => this._fwhmUnit;
+            set {
+                this._fwhmUnit = value;
+                this.RaisePropertyChanged();
+            }
+        }
+
+        public StarMeasurementUnit HFRStDevUnit {
+            get => this._hfrStDevUnit;
+            set {
+                this._hfrStDevUnit = value;
                 this.RaisePropertyChanged();
             }
         }