+++ NINA.WPF.Base/Model/ImageHistoryPoint.cs
@@ -14,6 +14,7 @@
 
 using CsvHelper.Configuration;
 using NINA.Image.ImageData;
+using NINA.Core.Enum;
 using NINA.Core.Utility;
 using NINA.Equipment.Interfaces.Mediator;
 using System;
@@ -72,7 +73,13 @@ namespace NINA.WPF.Base.Model {
                 return;
             if (imageSavedEventArgs.StarDetectionAnalysis != null) {
                 HFR = imageSavedEventArgs.StarDetectionAnalysis.HFR;
+                FWHM = imageSavedEventArgs.StarDetectionAnalysis.FWHM;
+                Eccentricity = imageSavedEventArgs.StarDetectionAnalysis.Eccentricity;
+                HFRUnit = imageSavedEventArgs.StarDetectionAnalysis.HFRUnit;
+                FWHMUnit = imageSavedEventArgs.StarDetectionAnalysis.FWHMUnit;
+                HFRStDevUnit = imageSavedEventArgs.StarDetectionAnalysis.HFRStDevUnit;
                 Stars = imageSavedEventArgs.StarDetectionAnalysis.DetectedStars;
+                RaiseStarMeasurementPropertyChanges();
             }
             IsBayered = imageSavedEventArgs.IsBayered;
             Filter = imageSavedEventArgs.Filter;
@@ -110,6 +117,14 @@ namespace NINA.WPF.Base.Model {
             }
         }
 
+        public void SetArcsecPerPixel(double value) {
+            arcsecPerPixel = value;
+            RaisePropertyChanged(nameof(HFRPixels));
+            RaisePropertyChanged(nameof(FWHMPixels));
+            RaisePropertyChanged(nameof(HFRArcseconds));
+            RaisePropertyChanged(nameof(FWHMArcseconds));
+        }
+
         public int Id { get; private set; }
         public int Index { get; set; }
         public double Zero { get; } = 0.05;
@@ -119,6 +134,15 @@ namespace NINA.WPF.Base.Model {
         public NINA.Core.Model.AutoFocusPoint AutoFocusPoint { get; private set; }
 
         public double HFR { get; private set; }
+        public double FWHM { get; private set; } = double.NaN;
+        public double Eccentricity { get; private set; } = double.NaN;
+        public StarMeasurementUnit HFRUnit { get; private set; } = StarMeasurementUnit.Pixels;
+        public StarMeasurementUnit FWHMUnit { get; private set; } = StarMeasurementUnit.Arcseconds;
+        public StarMeasurementUnit HFRStDevUnit { get; private set; } = StarMeasurementUnit.Pixels;
+        public double HFRPixels => ConvertStarMeasurement(HFR, HFRUnit, StarMeasurementUnit.Pixels);
+        public double FWHMPixels => ConvertStarMeasurement(FWHM, FWHMUnit, StarMeasurementUnit.Pixels);
+        public double HFRArcseconds => ConvertStarMeasurement(HFR, HFRUnit, StarMeasurementUnit.Arcseconds);
+        public double FWHMArcseconds => ConvertStarMeasurement(FWHM, FWHMUnit, StarMeasurementUnit.Arcseconds);
 
         public int Stars { get; private set; }
 
@@ -146,6 +170,27 @@ namespace NINA.WPF.Base.Model {
         public int FocuserPosition { get; private set; } = 0;
         public double RotatorPosition { get; private set; } = 0.0;
         public double RotatorMechanicalPosition { get; private set; } = 0.0;
+
+        private double arcsecPerPixel = double.NaN;
+
+        private double ConvertStarMeasurement(double value, StarMeasurementUnit sourceUnit, StarMeasurementUnit targetUnit) {
+            return StarMeasurementUnitConverter.TryConvert(value, sourceUnit, targetUnit, arcsecPerPixel, out var convertedValue)
+                ? convertedValue
+                : double.NaN;
+        }
+
+        private void RaiseStarMeasurementPropertyChanges() {
+            RaisePropertyChanged(nameof(HFR));
+            RaisePropertyChanged(nameof(FWHM));
+            RaisePropertyChanged(nameof(Eccentricity));
+            RaisePropertyChanged(nameof(HFRUnit));
+            RaisePropertyChanged(nameof(FWHMUnit));
+            RaisePropertyChanged(nameof(HFRStDevUnit));
+            RaisePropertyChanged(nameof(HFRPixels));
+            RaisePropertyChanged(nameof(FWHMPixels));
+            RaisePropertyChanged(nameof(HFRArcseconds));
+            RaisePropertyChanged(nameof(FWHMArcseconds));
+        }
     }
 
     public sealed class ImageHistoryPointMap : ClassMap<ImageHistoryPoint> {
@@ -179,6 +224,9 @@ namespace NINA.WPF.Base.Model {
             Map(m => m.AutoFocusPoint.NewPosition).Optional().Index(26).Name("AutoFocus New Position");
             Map(m => m.AutoFocusPoint.Time).Index(27).Optional().Name("AutoFocus Time");
             Map(m => m.AutoFocusPoint.Filter).Optional().Index(28).Name("AutoFocus Filter");
+            Map(m => m.FWHM).Optional().Index(29).Name("FWHM");
+            Map(m => m.HFRUnit).Optional().Index(30).Name("HFR Unit");
+            Map(m => m.FWHMUnit).Optional().Index(31).Name("FWHM Unit");
         }
     }
 }